reworking code into seperable modules
