+++ lib/Cache.js
@@ -53,8 +53,9 @@ cp.put = function (requestInfo, responsePromise) {
             }
         };
 
-        var onResponseError = function () {
-            requestToResponse.remove(requestInfo.key);
+        var onResponseError = function (e) {
+            console.log(e.stack);
+            requestToResponse.delete(requestInfo.key);
         };
 
         responsePromise.then(onResponse, onResponseError);
