+++ lib/utils.js
@@ -1,7 +1,6 @@
 var util = require('util');
-var EventEmitter = require('events').EventEmitter;
 var url = require('url');
-
+var Promise = require("bluebird");
 var parseUrl =  url.parse;
 
 var define = function (obj, prop, value) {
@@ -31,39 +30,11 @@ var getOrigin = function(u) {
     return u.protocol + '//' + u.host + ':' + resolvePort(u);
 };
 
-//////////////////////////////////////////      ConfEmitter           //////////////////////////////////////////
-
-/*
- * If configuration is taking place, wait on all requests
- */
-function ConfEmitter() {
-    this.activeConfigurationCnt = 0;
-    EventEmitter.call(this);
-}
-
-util.inherits(ConfEmitter, EventEmitter);
-
-ConfEmitter.prototype.increment = function () {
-    this.activeConfigurationCnt++;
-};
-
-ConfEmitter.prototype.decrement = function () {
-    this.activeConfigurationCnt--;
-    if (this.activeConfigurationCnt === 0) {
-        this.emit('completed');
-    }
-};
-
-ConfEmitter.prototype.configuring = function () {
-    return this.activeConfigurationCnt > 0;
-};
-
-
-//////////////////////////////////////////      Exports           //////////////////////////////////////////
+//////////////////////////////////////////      Exports                     //////////////////////////////////////////
 module.exports = {
     define: define,
     resolvePort: resolvePort,
     getOrigin: getOrigin,
-    ConfEmitter: ConfEmitter,
     parseUrl:  parseUrl
 };
+
